refactor: some slot refactoring
