+++ src/catalog/CatalogPage.tsx
@@ -8,6 +8,7 @@ import { DEFAULT_PAGE_SIZE } from '@src/data/course-list-search/constants';
 import { useCourseListSearch } from '@src/data/course-list-search/hooks';
 import CourseCatalogIntroSlot from '@src/plugin-slots/CourseCatalogIntroSlot';
 import { CourseCatalogDataTableSlot } from '@src/plugin-slots/CourseCatalogDataTableSlots';
+import CourseCatalogSearchFieldSlot from '@src/plugin-slots/CourseCatalogSearchFieldSlot';
 import { useDebouncedSearchInput } from './hooks/useDebouncedSearchInput';
 import { AlertNotification, Loading } from '../generic';
 import { useCatalog } from './hooks/useCatalog';
@@ -93,16 +94,17 @@ const CatalogPage = () => {
     <Container fluid={false} size="xl" className="pt-5.5 mb-6">
       <CourseCatalogIntroSlot searchString={searchString} courseDataResultsLength={courseData?.results?.length} />
       {hasCourses ? (
-        <CourseCatalogDataTableSlot
-          displayData={displayData}
-          totalCourses={totalCourses}
-          pageCount={pageCount}
-          pageIndex={pageIndex}
-          tableColumns={tableColumns}
-          handleFetchData={handleFetchData}
-          setSearchInput={setSearchInput}
-          handleSearch={handleSearch}
-        />
+        <>
+          <CourseCatalogSearchFieldSlot setSearchInput={setSearchInput} handleSearch={handleSearch} />
+          <CourseCatalogDataTableSlot
+            displayData={displayData}
+            totalCourses={totalCourses}
+            pageCount={pageCount}
+            pageIndex={pageIndex}
+            tableColumns={tableColumns}
+            handleFetchData={handleFetchData}
+          />
+        </>
       ) : (
         <AlertNotification
           title={intl.formatMessage(messages.noCoursesAvailable)}
