+++ src/plugin-slots/CourseCatalogDataTableSlots/CourseCatalogDataTableCardViewSlot/CourseCatalogDataTableCourseCardSlot/index.tsx
@@ -10,7 +10,7 @@ const CourseCatalogDataTableCourseCardSlot = ({ original: courseData, isLoading 
       mergeProps: true,
     }}
     pluginProps={{
-      original: courseData,
+      courseData,
       isLoading,
     }}
   >
