+++ src/plugin-slots/CourseCatalogSearchFieldSlot/images/screenshot_custom.png
