+++ src/plugin-slots/CourseCatalogSearchFieldSlot/images/screenshot_custom_with_form_control.png
