+++ src/plugin-slots/CourseCatalogSearchFieldSlot/images/screenshot_default.png
