+++ src/plugin-slots/CourseCatalogSearchFieldSlot/types.ts
@@ -1,4 +1,4 @@
-export interface CourseCatalogDataTableSearchFieldSlotProps {
+export interface CourseCatalogSearchFieldSlotProps {
   setSearchInput: (value: string) => void;
   handleSearch: (value: string) => void;
 }
